tests: move half precision tests to another file
